refactor: util method 분리
